requirements.txt
flask==3.0.3
flask-cors==4.0.1
playwright==1.44.0
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.32.3
